third_party/WebKit/LayoutTests/fast/dom/Window/orphaned-frame-access.html

<!DOCTYPE HTML>
<html>
<head>
<title>Null frame access tests</title>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script>
var t = async_test('This tests access to a window with a null frame.');

window.onload = t.step_func(function () {
  var childWindow = window[0];
  childWindow.frameElement.remove();
  // Removing the owner element from the DOM should null the frame. In Blink,
  // this can be observed by checking childWindow.parent == null.
  assert_equals(childWindow.parent, null);

  assert_equals(childWindow.test, 'abc', 'Named property should still be accessible.');
  assert_equals(childWindow[20], '123', 'Indexed property should still be accessible.');
  // TODO(dcheng): Why does this return null?
  assert_equals(childWindow.Comment, undefined, 'Interface Object should be gone.');
  t.done();
});
</script>
</head>
<body>
<iframe id="subframe" src="data:text/html,&lt;script&gt;window.test='abc';window[20]='123';&lt;/script&gt;" />
</body>
</html>
